+++ aderyn_core/src/lib.rs
@@ -13,10 +13,17 @@ use std::path::{Path, PathBuf};
 
 use crate::context::loader::ContextLoader;
 use crate::detect::detector::{get_all_detectors, IssueSeverity};
-use crate::report::printer::{MarkdownReportPrinter, ReportPrinter};
+use crate::report::printer::ReportPrinter;
 use crate::report::reporter::{Issue, Report};
 
-pub fn run(context_loader: ContextLoader, output_file_path: String) -> Result<(), Box<dyn Error>> {
+pub fn run_with_printer<T>(
+    context_loader: ContextLoader,
+    output_file_path: String,
+    reporter: T,
+) -> Result<(), Box<dyn Error>>
+where
+    T: ReportPrinter<()>,
+{
     println!("Get Detectors");
 
     let detectors = get_all_detectors();
@@ -55,9 +62,8 @@ pub fn run(context_loader: ContextLoader, output_file_path: String) -> Result<()
 
     println!("Detectors run, processing found issues");
 
-    let printer = MarkdownReportPrinter;
     println!("Found issues processed. Printing report");
-    printer.print_report(
+    reporter.print_report(
         get_markdown_writer(&output_file_path)?,
         &report,
         &context_loader,
