+++ aderyn_core/src/report/mod.rs
@@ -1,2 +1,4 @@
+pub mod json_printer;
+pub mod markdown_printer;
 pub mod printer;
 pub mod reporter;
